refactor: merge combined & hyper report renderer
