+++ optimize/client/src/modules/components/ReportRenderer/HyperReportRenderer.test.js
@@ -9,71 +9,88 @@
 import React from 'react';
 import {shallow} from 'enzyme';
 
+import {processResult} from 'services';
+
+import {Chart} from './visualizations';
 import HyperReportRenderer from './HyperReportRenderer';
+import ProcessReportRenderer from './ProcessReportRenderer';
+
+jest.mock('./service', () => {
+  return {
+    getFormatter: (view) => (v) => v,
+  };
+});
+
+jest.mock('services', () => {
+  const rest = jest.requireActual('services');
+  return {
+    ...rest,
+    formatters: {formatReportResult: (data, result) => result},
+    processResult: jest.fn().mockImplementation(({result}) => result),
+  };
+});
+
+const userTaskAssigneeReport = {
+  combined: false,
+  reportType: 'process',
+  data: {
+    processDefinitionKey: 'aKey',
+    processDefinitionVersion: '1',
+    view: {
+      properties: ['duration'],
+      entity: 'userTask',
+    },
+    groupBy: {
+      type: 'assignee',
+    },
+    visualization: 'bar',
+    distributedBy: {type: 'userTask', value: null},
+    configuration: {},
+  },
+  result: {
+    measures: [
+      {
+        property: 'duration',
+        data: [
+          {
+            key: 'Anne',
+            label: 'Anne',
+            value: [
+              {key: 'taskId1', label: 'Usertask 1', value: 8},
+              {key: 'taskId2', label: 'Usertask 2', value: 1},
+              {key: 'taskId3', label: 'Usertask 3', value: 65},
+            ],
+          },
+          {
+            key: 'Bernd',
+            label: 'Bernd',
+            value: [
+              {key: 'taskId1', label: 'Usertask 1', value: 3},
+              {key: 'taskId2', label: 'Usertask 2', value: 17},
+              {key: 'taskId3', label: 'Usertask 3', value: 22},
+            ],
+          },
+          {
+            key: 'Chris',
+            label: 'Chris',
+            value: [
+              {key: 'taskId1', label: 'Usertask 1', value: 1},
+              {key: 'taskId2', label: 'Usertask 2', value: 0},
+              {key: 'taskId3', label: 'Usertask 3', value: 73},
+            ],
+          },
+        ],
+      },
+    ],
+    instanceCount: 1234,
+    type: 'hyperMap',
+  },
+};
 
 it('should convert a hypermap to a combined report', () => {
-  const node = shallow(
-    <HyperReportRenderer
-      report={{
-        combined: false,
-        reportType: 'process',
-        data: {
-          processDefinitionKey: 'aKey',
-          processDefinitionVersion: '1',
-          view: {
-            properties: ['duration'],
-            entity: 'userTask',
-          },
-          groupBy: {
-            type: 'assignee',
-          },
-          visualization: 'bar',
-          distributedBy: {type: 'userTask', value: null},
-          configuration: {},
-        },
-        result: {
-          measures: [
-            {
-              property: 'duration',
-              data: [
-                {
-                  key: 'Anne',
-                  label: 'Anne',
-                  value: [
-                    {key: 'taskId1', label: 'Usertask 1', value: 8},
-                    {key: 'taskId2', label: 'Usertask 2', value: 1},
-                    {key: 'taskId3', label: 'Usertask 3', value: 65},
-                  ],
-                },
-                {
-                  key: 'Bernd',
-                  label: 'Bernd',
-                  value: [
-                    {key: 'taskId1', label: 'Usertask 1', value: 3},
-                    {key: 'taskId2', label: 'Usertask 2', value: 17},
-                    {key: 'taskId3', label: 'Usertask 3', value: 22},
-                  ],
-                },
-                {
-                  key: 'Chris',
-                  label: 'Chris',
-                  value: [
-                    {key: 'taskId1', label: 'Usertask 1', value: 1},
-                    {key: 'taskId2', label: 'Usertask 2', value: 0},
-                    {key: 'taskId3', label: 'Usertask 3', value: 73},
-                  ],
-                },
-              ],
-            },
-          ],
-          instanceCount: 1234,
-          type: 'hyperMap',
-        },
-      }}
-    />
-  );
+  const node = shallow(<HyperReportRenderer report={userTaskAssigneeReport} />);
 
-  expect(node).toMatchSnapshot();
+  expect(node.find(Chart).prop('report').combined).toBe(true);
 });
 
 it('should render single process report if report result is empty', () => {
@@ -89,66 +106,28 @@ it('should render single process report if report result is empty', () => {
     />
   );
 
-  expect(node).toMatchSnapshot();
+  expect(node.find(ProcessReportRenderer)).toExist();
 });
 
-it('should format single reports names for hyper reports distributed by date variable', () => {
-  const node = shallow(
-    <HyperReportRenderer
-      report={{
-        combined: false,
-        reportType: 'process',
-        data: {
-          processDefinitionKey: 'aKey',
-          processDefinitionVersion: '1',
-          view: {
-            properties: ['frequency'],
-            entity: 'processInstance',
-          },
-          groupBy: {
-            type: 'startDate',
-          },
-          visualization: 'bar',
-          distributedBy: {
-            type: 'variable',
-            value: {type: 'Date'},
-          },
-          configuration: {
-            distributeByDateVariableUnit: 'month',
-          },
-        },
-        result: {
-          measures: [
-            {
-              property: 'frequency',
-              data: [
-                {
-                  key: '2020-09-01T00:00:00.000+0200',
-                  label: '2020-09-01T00:00:00.000+0200',
-                  value: [
-                    {
-                      key: '1969-12-07T14:52:00Z',
-                      value: 1.0,
-                      label: '1969-12-07T14:52:00Z',
-                    },
-                    {
-                      key: '2020-03-08T14:52:00Z',
-                      value: 1.0,
-                      label: '2020-03-08T14:52:00Z',
-                    },
-                  ],
-                },
-              ],
-            },
-          ],
-          instanceCount: 1234,
-          type: 'hyperMap',
-        },
-      }}
-    />
-  );
+it('should convert adjust the result to be identical to combined report result structure', () => {
+  const node = shallow(<HyperReportRenderer report={userTaskAssigneeReport} />);
 
-  const names = Object.values(node.prop('report').result.data).map(({name}) => name);
+  const resultData = node.find(Chart).prop('report').result.data;
 
-  expect(names).toEqual(['Dec 1969', 'Mar 2020']);
+  expect(resultData['taskId1'].result.data).toEqual(userTaskAssigneeReport.result.measures[0].data);
+  expect(resultData['taskId1'].result.measures[0].data).toEqual([
+    {key: 'Anne', label: 'Anne', value: 8},
+    {key: 'Bernd', label: 'Bernd', value: 3},
+    {key: 'Chris', label: 'Chris', value: 1},
+  ]);
+});
+
+it('should process the result of every created report', () => {
+  processResult.mockClear();
+
+  shallow(<HyperReportRenderer report={userTaskAssigneeReport} />);
+
+  expect(processResult).toHaveBeenCalledTimes(3);
+  const reportNames = processResult.mock.calls.map((call) => call[0].id);
+  expect(reportNames).toEqual(['taskId1', 'taskId2', 'taskId3']);
 });
